moved changes to tensorflow folder
